feat: add protecting the categories relations

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -96,6 +96,12 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+
+        if ($subCategory > 0) {
+            return response(['status' => 'error', 'message' => 'This items contains sub items for delete this you have to
+            delete the sub items first!']);
+        }
         $category->delete();
 
         return response(['status' => 'success', 'Deleted Successfully!']);
