+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -98,6 +98,12 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = SubCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if ($childCategory > 0) {
+            return response(['status' => 'error', 'message' => 'This item contains, sub items for delete this you have to
+            delete the sub items first!']);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully1']);
